refactor(utils): align utility layout with AGENTS structure
- move utilities to utils/<utility>/index.js and co-located index.test.js

- update route and handler imports to new paths

- no functional changes intended
