+++ utils/buildSearchSessionPreference/index.js
